Desktop/projects/erp2-erp_base-erp-backend/backend/src/requirements.txt
fastapi
uvicorn
aioredis==2.0.1
passlib
python-jose
sqlalchemy
pymysql
pika
starlette
pydantic